services/trade_producer/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["max-mcd <max-mcd@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
quixstreams = "^2.5.1"
websocket-client = "^1.8.0"
logger = "^1.4"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.3.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["I"]